requirements.txt
pandas
numpy
altair
streamlit
streamlit_lottie
Pillow